requirements.txt
astral==3.2
httpx==0.28.1
Jinja2==3.1.6
numpy==1.26.1
ollama==0.5.2
pandas==2.3.1
ppi-python==0.2.3
pydantic==2.11.7
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.14.1
